requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-extensions==3.2.3